+++ CHANGELOG.md
@@ -21,6 +21,9 @@
 ### Removed
 
 ### Fixed
+- SaveSelection:
+    - If mapMode is 3D it adds the mapMode, tilt, heading and altitude attribute to the url.
+    - In 3D the visible layer list contains the 3D layer now.
 
 ---